+++ Snoop.Core/Views/DiagnosticsView.xaml.cs
@@ -1,7 +1,6 @@
 ﻿namespace Snoop.Views;
 
 using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Collections.Specialized;
 using System.ComponentModel;
@@ -11,9 +10,7 @@ using System.Text;
 using System.Windows;
 using System.Windows.Data;
 using System.Windows.Input;
-using System.Xml.Serialization;
-using JetBrains.Annotations;
-using Snoop.Data.Tree;
+using Snoop.Infrastructure;
 using Snoop.Infrastructure.Diagnostics;
 using Snoop.Infrastructure.Helpers;
 
@@ -259,15 +256,11 @@ public partial class DiagnosticsView
 
         try
         {
-            var diagnosticItems = this.DiagnosticsItemsView.Cast<DiagnosticItem>()
-                .Select(x => new DiagnosticItemData(x.Area, x.Level, x.Name, x.Description, x.SourceObject?.ToString(), TreeItemDiagnosticPathData.Build(x.TreeItem)))
-                .ToList();
-
             var filePath = ExportHelper.GetUniqueExportFilePath("Diagnostics", "xml");
 
             using (var streamWriter = new StreamWriter(filePath, false, Encoding.UTF8))
             {
-                new XmlSerializer(typeof(List<DiagnosticItemData>)).Serialize(streamWriter, diagnosticItems);
+                DiagnosticsExporter.Export(this.DiagnosticsItemsView.Cast<DiagnosticItem>(), streamWriter);
             }
 
             MessageBox.Show($"The data has been exported to \"{filePath}\".", "Data exported", MessageBoxButton.OK, MessageBoxImage.Information);
@@ -275,60 +268,6 @@ public partial class DiagnosticsView
         finally
         {
             Mouse.OverrideCursor = saveCursor;
-        }
-    }
-
-    [PublicAPI]
-    public record DiagnosticItemData(DiagnosticArea Area, DiagnosticLevel Level, string Name, string Description, string? SourceObject, List<TreeItemDiagnosticPathData>? TreeItemPath)
-    {
-        public DiagnosticItemData()
-            : this(DiagnosticArea.Misc, DiagnosticLevel.Info, string.Empty, string.Empty, null, null)
-        {
-        }
-
-        public DiagnosticArea Area { get; set; } = Area;
-
-        public DiagnosticLevel Level { get; set; } = Level;
-
-        public string Name { get; set; } = Name;
-
-        public string Description { get; set; } = Description;
-
-        public string? SourceObject { get; set; } = SourceObject;
-
-        public List<TreeItemDiagnosticPathData>? TreeItemPath { get; set; } = TreeItemPath;
-    }
-
-    [PublicAPI]
-    public record TreeItemDiagnosticPathData(string DisplayName, string TypeName)
-    {
-        public TreeItemDiagnosticPathData()
-            : this(string.Empty, string.Empty)
-        {
-        }
-
-        public string DisplayName { get; set; } = DisplayName;
-
-        public string TypeName { get; set; } = TypeName;
-
-        public static List<TreeItemDiagnosticPathData>? Build(TreeItem? treeItem)
-        {
-            if (treeItem is null)
-            {
-                return null;
-            }
-
-            var items = new List<TreeItemDiagnosticPathData>();
-
-            var current = treeItem;
-            while (current is not null)
-            {
-                items.Add(new(current.DisplayName, current.TargetType.Name));
-                current = current.Parent;
-            }
-
-            items.Reverse();
-            return items;
         }
     }
 }
